Refactored server into module folders
