+++ server/public/index.html
@@ -191,7 +191,7 @@
 		<script src='node_modules/bootstrap/dist/js/bootstrap.min.js'></script>
 		<script src='node_modules/dropzone/dist/dropzone.js'></script>
 		<script src="lib/underscore.js"></script>
-		<script src='clip.js'></script>
+		<script src='modules/clip/clip.js'></script>
 		<script src="characterData.js"></script>
 		<script src='script.js'></script>
 	</body>
