+++ server/public/modules/clip/clip.js
