+++ server/public/modules/clip/endpoints.js
@@ -1,0 +1,3 @@
+module.exports = function(server){
+
+}
