+++ server/public/modules/youtube/endpoints.js
@@ -1,0 +1,47 @@
+'use strict';
+
+let Youtube = require('./youtube');
+
+module.exports = function(server){
+    server.get('/youtube/auth', function(req, res){
+        Youtube.oauth();
+    });
+
+    server.get('/youtube/authenticate', function(req, res){
+        var code = req.query.code;
+        var verify = Youtube.verifyOAuth(code);
+    });
+
+    server.get('/uploadStatus', function(req, res){
+        var id = req.query.id;
+        var uploadStatus = Youtube.getUploadStatus(id);
+        res.send(uploadStatus);
+    });
+
+    server.post('/upload', function(req, res){
+        var file = req.body.file;
+
+        var input 		= file.inputFileName;
+        var tournament 	= file.tournamentName;
+        var round 		= file.round;
+        var p1name 		= file.player1;
+        var p2name 		= file.player2;
+        var output 		= file.outputFileName;
+        var bracket 	= file.bracketUrl;
+
+        var yt = new Youtube(output, p1name, p2name, tournament, round, bracket);
+        if(!Youtube.isAuthenticated())
+            res.sendStatus(500);
+        else{
+            yt.upload()
+                .then(function(){
+                    res.sendStatus(200);
+                    res.end();
+                })
+                .catch(function(err){
+                    log.error(err.stack);
+                    res.sendStatus(500);
+                })
+        }
+    });
+}
