+++ server/public/modules/youtube/uploadManager.js
